feat: add generateSbom Task / apply cyclonDx Plugin
