+++ src/integrationTest/kotlin/com/liftric/dtcp/GenerateVexTest.kt
@@ -13,31 +13,9 @@ import java.nio.file.Paths
 import java.nio.file.StandardCopyOption
 import org.cyclonedx.parsers.JsonParser
 
-class GenerateVexTest {
-
+class GenerateVexTest: IntegrationTestBase() {
     @Test
     fun testGenerateVexTask() {
-
-        val vexComponent = VexComponent(
-            purl = "pkg:maven/org.eclipse.jetty/jetty-http@9.4.49.v20220914?type=jar",
-            vulnerability = VexVulnerability(
-                id = "CVE-2023-26048",
-                source = "NVD",
-                analysis = "Vulnerability.Analysis.State.FALSE_POSITIVE",
-                analysisValue = "FALSE_POSITIVE",
-                detail = null,
-            )
-        )
-
-        val vexVulnerability = VexVulnerability(
-            id = "CVE-2020-8908",
-            source = "NVD",
-            analysis = "Vulnerability.Analysis.State.RESOLVED",
-            analysisValue = "RESOLVED",
-            detail = "This is resolved",
-        )
-
-
         val projectDir = File("build/generateVexTest")
 
         projectDir.mkdirs()
@@ -48,7 +26,12 @@ import com.liftric.dtcp.extensions.*
 import org.cyclonedx.model.vulnerability.Vulnerability
 
 plugins {
+    kotlin("jvm") version "1.8.21"
     id("com.liftric.dependency-track-companion-plugin")
+}
+
+repositories {
+    mavenCentral()
 }
 
 group = "com.liftric.test"
@@ -72,11 +55,6 @@ dependencyTrackCompanion {
 }
         """
         )
-
-        val sourceJsonFile = Paths.get("test/data/bom.json")
-        val targetJsonFile = projectDir.toPath().resolve("build/reports/bom.json")
-        Files.createDirectories(targetJsonFile.parent)
-        Files.copy(sourceJsonFile, targetJsonFile, StandardCopyOption.REPLACE_EXISTING)
 
         val result = GradleRunner
             .create()
